requirements.txt
streamlit
langchain
langchain-community
openai
python-dotenv
pypdf
tiktoken
faiss-cpu
langchain-openai
